backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
sqlalchemy<2.0
pydantic==1.10.13
python-dotenv==1.0.0
passlib[bcrypt]
python-jose[cryptography]